moved Dlib Tracker to CameraWidget, crashes with more than one camera is tracking
