+++ views/widgets/camera_widget.py
@@ -6,6 +6,7 @@ from PySide6.QtCore import Signal
 import cv2
 import time
 import imutils
+import dlib
 import shared.constants as constants
 from logic.facial_tracking.testing_image_processor import ImageProcessor
 from views.widgets.video_thread import VideoThread
@@ -46,6 +47,15 @@ class CameraWidget(QLabel):
         self.processor_thread = ImageProcessor(stream_thread=self.stream_thread)
         self.processor_thread.start()
 
+        self.tracker = None
+        self.temp_tracked_name = None
+        self.track_x = None
+        self.track_y = None
+        self.track_w = None
+        self.track_h = None
+        self.is_tracking = False  # If Track Checkbox is checked
+        self.tracked_name = None  # Face that needs to be tracked
+
     def stop(self):
         """
         When CameraWidget is being removed from the UI, we should stop all relevant threads before deletion.
@@ -84,29 +94,22 @@ class CameraWidget(QLabel):
             self.processor_thread.start()
 
     def set_tracking(self):
-        if self.processor_thread.is_alive():
-            self.processor_thread.is_tracking = not self.processor_thread.is_tracking
+        self.is_tracking = not self.is_tracking
 
     def set_tracked_name(self, name):
-        print(f"Setting {name} as tracked name")
-        if self.processor_thread.is_alive():
-            self.processor_thread.tracked_name = name
+        self.tracked_name = name
 
     def get_tracked_name(self):
-        if self.processor_thread.is_alive():
-            return self.processor_thread.tracked_name
+        return self.tracked_name
 
     def get_tracking(self):
-        if self.processor_thread.is_alive():
-            return self.processor_thread.is_tracking
+        return self.is_tracking
 
     def update_image(self, cv_img):
         """Updates the QLabel with the latest OpenCV/NDI frame and draws it"""
         cv_img = self.draw_on_frame(frame=cv_img, face_locations=self.processor_thread.face_locations,
                                     face_names=self.processor_thread.face_names,
-                                    confidence_list=self.processor_thread.confidence_list,
-                                    track_x=self.processor_thread.track_x, track_y=self.processor_thread.track_y,
-                                    track_w=self.processor_thread.track_w, track_h=self.processor_thread.track_h)
+                                    confidence_list=self.processor_thread.confidence_list)
         qt_img = self.convert_cv_qt(cv_img)
         self.setPixmap(qt_img)
 
@@ -145,7 +148,7 @@ class CameraWidget(QLabel):
             constants.CURRENT_ACTIVE_CAM_WIDGET.update()
         self.change_selection_signal.emit()
 
-    def draw_on_frame(self, frame, face_locations, face_names, confidence_list, track_x, track_y, track_w, track_h):
+    def draw_on_frame(self, frame, face_locations, face_names, confidence_list):
         """
         Is called by update_image and returns the latest frame with FPS + face box drawings if there are any.
         :param frame:
@@ -161,22 +164,21 @@ class CameraWidget(QLabel):
                 right *= 2
                 bottom *= 2
                 left *= 2
-                # print(top, right, bottom, left)
+
+                if name == self.tracked_name:
+                    self.temp_tracked_name = name
+                    self.track_x = left
+                    self.track_y = top
+                    self.track_w = right
+                    self.track_h = bottom
                 # Draw a box around the face
                 cv2.rectangle(frame, (left, top), (right, bottom), (0, 255, 0), 2)
                 # Draw a label with name and confidence for the face
                 cv2.putText(frame, name, (left + 5, top - 5), constants.FONT, 0.5, (255, 255, 255), 1)
                 cv2.putText(frame, confidence, (right - 52, bottom - 5), constants.FONT, 0.45, (255, 255, 0), 1)
 
-        if self.get_tracking():
-            # min_x = int(frame.shape[1] / 11.5)
-            # max_x = int(frame.shape[1] / 1.1)
-            # min_y = int(frame.shape[0] / 8.5)
-            # max_y = int(frame.shape[0] / 1.3)
-            # cv2.rectangle(frame, (min_x, min_y), (max_x, max_y), (255, 0, 0), 2)
-            if track_x is not None:
-                cv2.putText(frame, "tracking", (track_x, track_h + 15), constants.FONT, 0.45, (0, 255, 0), 1)
-                cv2.rectangle(frame, (track_x, track_y), (track_w, track_h), (255, 0, 255), 3, 1)
+        if self.is_tracking and self.processor_thread.is_alive():
+            frame = self.track_face(frame)
 
         # FPS Counter
         self.fc += 1
@@ -188,6 +190,37 @@ class CameraWidget(QLabel):
         fps = "FPS: " + str(self.FPS)[:5]
 
         cv2.putText(frame, fps, (50, 50), constants.FONT, 1, (0, 0, 255), 2)
+        return frame
+
+    def track_face(self, frame):  # Probably needs to be on its own thread
+        rgb_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+        if self.tracker is None and self.track_x is not None:
+            self.tracker = dlib.correlation_tracker()
+            rect = dlib.rectangle(self.track_x, self.track_y, self.track_w, self.track_h)
+            self.tracker.start_track(rgb_frame, rect)
+            cv2.putText(frame, "tracking", (self.track_x, self.track_h + 15), constants.FONT, 0.45, (0, 255, 0), 1)
+            cv2.rectangle(frame, (self.track_x, self.track_y), (self.track_w, self.track_h), (255, 0, 255), 3, 1)
+        if self.tracker is not None and self.temp_tracked_name == self.tracked_name and self.track_x is not None:
+            rect = dlib.rectangle(self.track_x, self.track_y, self.track_w, self.track_h)
+            self.tracker.start_track(rgb_frame, rect)
+            cv2.putText(frame, "tracking", (self.track_x, self.track_h + 15), constants.FONT, 0.45, (0, 255, 0), 1)
+            cv2.rectangle(frame, (self.track_x, self.track_y), (self.track_w, self.track_h), (255, 0, 255), 3, 1)
+        elif self.tracker is not None:
+            self.tracker.update(rgb_frame)
+            pos = self.tracker.get_position()
+            # unpack the position object
+            self.track_x = int(pos.left())
+            self.track_y = int(pos.top())
+            self.track_w = int(pos.right())
+            self.track_h = int(pos.bottom())
+            cv2.putText(frame, "tracking", (self.track_x, self.track_h + 15), constants.FONT, 0.45, (0, 255, 0), 1)
+            cv2.rectangle(frame, (self.track_x, self.track_y), (self.track_w, self.track_h), (255, 0, 255), 3, 1)
+
+        self.track_x = None
+        self.track_y = None
+        self.track_w = None
+        self.track_h = None
+        self.temp_tracked_name = None
         return frame
 
     def closeEvent(self, event):
